+++ core/helpers.py
@@ -1,9 +1,16 @@
 from collections import namedtuple
 from urllib.parse import urlencode
 from django.urls import reverse
+import collections
+
+import directory_components.helpers
+from directory_api_client.client import api_client
+from django.http import Http404
 from django.utils import translation
 import requests
 
+from directory_constants import choices
+from django.utils.html import escape, mark_safe
 from core import constants
 
 
@@ -23,6 +30,132 @@ def get_language_from_querystring(request):
     language_code = request.GET.get('language') or request.GET.get('lang')
     if language_code and language_code in language_codes:
         return language_code
+
+
+GA_DATA_MAPPING = {
+    'InternationalHomePage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'HomePage',
+        'site_subsection': ''
+    },
+    'InvestInternationalHomePage': {
+        'business_unit': 'Invest',
+        'site_section': 'HomePage',
+        'site_subsection': ''
+    },
+    'InvestHighPotentialOpportunityDetailPage': {
+        'business_unit': 'Invest',
+        'site_section': 'HighPotentialOpportunities',
+        'site_subsection': 'DetailPage',
+    },
+    'InternationalTopicLandingPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'Topic',
+        'site_subsection': 'ListingPage'
+    },
+    'InternationalArticleListingPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'Article',
+        'site_subsection': 'ListingPage'
+    },
+    'InternationalArticlePage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'Article',
+        'site_subsection': 'DetailPage'
+    },
+    'InternationalCampaignPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'Campaign',
+        'site_subsection': 'LandingPage'
+    },
+    'InternationalSectorPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'Sector',
+        'site_subsection': 'DetailPage'
+    },
+    'InternationalSubSectorPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'SubSector',
+        'site_subsection': 'DetailPage'
+    },
+    'InternationalCuratedTopicLandingPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'CuratedTopic',
+        'site_subsection': 'LandingPage'
+    },
+    'InternationalGuideLandingPage': {
+        'business_unit': 'Invest',
+        'site_section': 'Guide',
+        'site_subsection': 'ListingPage'
+    },
+    'InternationalEUExitFormPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'EUExit',
+        'site_subsection': 'FormPage'
+    },
+    'InternationalEUExitFormSuccessPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'EUExit',
+        'site_subsection': 'FormSuccessPage'
+    },
+    'InternationalCapitalInvestLandingPage': {
+        'business_unit': 'CapitalInvestment',
+        'site_section': 'LandingPage',
+        'site_subsection': ''
+    },
+    'AboutDitServicesPage': {
+        'business_unit': 'GreatInternational',
+        'site_section': 'AboutDIT',
+        'site_subsection': 'ServicesPage'
+    },
+    'CapitalInvestRegionPage': {
+        'business_unit': 'CapitalInvestment',
+        'site_section': 'Region',
+        'site_subsection': 'DetailPage'
+    },
+    'CapitalInvestOpportunityPage': {
+        'business_unit': 'CapitalInvestment',
+        'site_section': 'Opportunity',
+        'site_subsection': 'DetailPage'
+    },
+    'CapitalInvestOpportunityListingPage': {
+        'business_unit': 'CapitalInvestment',
+        'site_section': 'Opportunity',
+        'site_subsection': 'ListingPage'
+    },
+}
+
+HEADER_MAPPING = {
+    'InternationalHomePage': '',
+    'InvestInternationalHomePage': 'invest',
+    'InvestHighPotentialOpportunityDetailPage': 'invest',
+    'InternationalTopicLandingPage': 'industries',
+    'InternationalArticleListingPage': 'uk_setup_guide',
+    'InternationalArticlePage': 'uk_setup_guide',
+    'InternationalCampaignPage': '',
+    'InternationalSectorPage': 'industries',
+    'InternationalSubSectorPage': 'industries',
+    'InternationalCuratedTopicLandingPage': 'uk_setup_guide',
+    'InternationalGuideLandingPage': 'uk_setup_guide',
+    'InternationalEUExitFormPage': '',
+    'InternationalEUExitFormSuccessPage': '',
+    'InternationalCapitalInvestLandingPage': 'invest',
+    'AboutDitServicesPage': '',
+    'CapitalInvestRegionPage': 'invest',
+    'CapitalInvestOpportunityPage': 'invest',
+    'CapitalInvestOpportunityListingPage': 'invest'
+}
+
+
+NotifySettings = collections.namedtuple(
+    'NotifySettings',
+    [
+        'company_template',
+        'support_template',
+        'investor_template',
+        'support_email_address',
+    ]
+)
 
 
 def get_ga_data_for_page(page_type):
@@ -148,6 +281,88 @@ def sort_opportunities(opportunities, sort_by_chosen):
         )
 
     return opportunities
+
+
+class CompanyParser(directory_components.helpers.CompanyParser):
+
+    def serialize_for_template(self):
+        if not self.data:
+            return {}
+        return {
+            **self.data,
+            'date_of_creation': self.date_of_creation,
+            'address': self.address,
+            'sectors': self.sectors_label,
+            'keywords': self.keywords,
+            'employees': self.employees_label,
+            'expertise_industries': self.expertise_industries_label,
+            'expertise_regions': self.expertise_regions_label,
+            'expertise_countries': self.expertise_countries_label,
+            'expertise_languages': self.expertise_languages_label,
+            'has_expertise': self.has_expertise,
+            'expertise_products_services': (
+                self.expertise_products_services_label
+            ),
+            'is_in_companies_house': self.is_in_companies_house,
+        }
+
+
+def get_results_from_search_response(response):
+    parsed = response.json()
+    formatted_results = []
+
+    for result in parsed['hits']['hits']:
+        parser = CompanyParser(result['_source'])
+        formatted = parser.serialize_for_template()
+        if 'highlight' in result:
+            highlighted = '...'.join(
+                result['highlight'].get('description', '') or
+                result['highlight'].get('summary', '')
+            )
+            # escape all html tags other than <em> and </em>
+            highlighted_escaped = (
+                escape(highlighted)
+                .replace('&lt;em&gt;', '<em>')
+                .replace('&lt;/em&gt;', '</em>')
+            )
+            formatted['highlight'] = mark_safe(highlighted_escaped)
+        formatted_results.append(formatted)
+
+    parsed['results'] = formatted_results
+    return parsed
+
+
+def get_filters_labels(filters):
+    sectors = dict(choices.INDUSTRIES)
+    languages = dict(choices.EXPERTISE_LANGUAGES)
+    labels = []
+    skip_fields = [
+        'q',
+        'page',
+        # Prevents duplicates labels not to be displayed in filter list
+        'expertise_products_services_label'
+    ]
+    for name, values in filters.items():
+        if name in skip_fields:
+            pass
+        elif name == 'industries':
+            labels += [sectors[item] for item in values if item in sectors]
+        elif name == 'expertise_languages':
+            labels += [languages[item] for item in values if item in languages]
+        elif name.startswith('expertise_products_services_'):
+            labels += values
+        else:
+            for value in values:
+                labels.append(value.replace('_', ' ').title())
+    return labels
+
+
+def get_company_profile(number):
+    response = api_client.company.retrieve_public_profile(number=number)
+    if response.status_code == 404:
+        raise Http404(f'API returned 404 for company number {number}')
+    response.raise_for_status()
+    return response.json()
 
 
 def count_data_with_field(list_of_data, field):