+++ core/mixins.py
@@ -13,7 +13,42 @@ from directory_constants import cms
 from directory_cms_client.client import cms_api_client
 from directory_cms_client.helpers import handle_cms_response
 
-from core import constants
+from core import constants, helpers
+
+TEMPLATE_MAPPING = {
+    'InternationalHomePage': 'core/landing_page.html',
+    'InternationalTopicLandingPage': 'core/topic_list.html',
+    'InternationalArticleListingPage': 'core/article_list.html',
+    'InternationalArticlePage': 'core/uk_setup_guide/article_detail.html',
+    'InternationalCampaignPage': 'core/campaign.html',
+    'InternationalSectorPage': 'core/sector_page.html',
+    'InternationalSubSectorPage': 'core/sector_page.html',
+    'InternationalCuratedTopicLandingPage': (
+        'core/how_to_do_business_landing_page.html'),
+    'InternationalGuideLandingPage': (
+        'core/uk_setup_guide/guide_landing_page.html'),
+    'InternationalEUExitFormPage': 'euexit/international-contact-form.html',
+    'InternationalEUExitFormSuccessPage': (
+        'euexit/international-contact-form-success.html'),
+    'InternationalCapitalInvestLandingPage': (
+        'core/capital_invest/capital_invest_landing_page.html'),
+    'CapitalInvestRegionPage': (
+        'core/capital_invest/capital_invest_region_page.html'),
+    'CapitalInvestOpportunityPage': (
+        'core/capital_invest/capital_invest_opportunity_page.html'),
+    'CapitalInvestOpportunityListingPage': (
+        'core/capital_invest/capital_invest_opportunity_listing_page.html'),
+    'AboutDitServicesPage': 'core/about_dit/services_page.html',
+    # Invest
+    'InvestInternationalHomePage': 'invest/landing_page.html',
+    'InvestHighPotentialOpportunityDetailPage': (
+        'invest/high_potential_opportunity_detail.html'),
+}
+
+FEATURE_FLAGGED_URLS_MAPPING = {
+    '/international/content/how-to-do-business-with-the-uk/': (
+        'HOW_TO_DO_BUSINESS_ON'),
+}
 
 
 class NotFoundOnDisabledFeature:
@@ -93,3 +128,42 @@ class SetEtagMixin:
         if request.method == 'GET':
             response.add_post_render_callback(set_response_etag)
         return response
+
+
+class SubmitFormOnGetMixin:
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+        data = self.request.GET or {}
+        if data:
+            kwargs['data'] = data
+        return kwargs
+
+    def get(self, request, *args, **kwargs):
+        return super().post(request, *args, **kwargs)
+
+
+class CompanyProfileMixin:
+    @cached_property
+    def company(self):
+        return helpers.get_company_profile(self.kwargs['company_number'])
+
+    def get_context_data(self, **kwargs):
+        company = helpers.CompanyParser(self.company)
+        return super().get_context_data(
+            company=company.serialize_for_template(),
+            **kwargs
+        )
+
+
+class PersistSearchQuerystringMixin:
+
+    @property
+    def search_querystring(self):
+        return self.request.GET.urlencode()
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(
+            search_querystring=self.search_querystring,
+            **kwargs,
+        )